requirements.txt
altair==5.4.1
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dacite==1.8.1
fonttools==4.54.1
gitdb==4.0.11
GitPython==3.1.41
htmlmin==0.1.12
idna==3.10
ImageHash==4.3.1
Jinja2==3.1.4
joblib==1.1.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
missingno==0.5.2
multimethod==1.12
narwhals==1.9.1
networkx==3.4.1
numba==0.60.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pandas-profiling==3.2.0
patsy==0.5.6
phik==0.12.4
pillow==11.0.0
protobuf==5.28.2
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.7.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
statsmodels==0.14.4
streamlit==1.39.0
streamlit-pandas-profiling==0.1.3
tangled-up-in-unicode==0.2.0
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
visions==0.7.4
watchdog==5.0.3
wordcloud==1.9.3
ydata-profiling==4.11.0
